pyproject.toml
[tool.poetry]
name = "github-collector"
version = "0.1.0"
description = ""
authors = ["nasum <tomato.wonder.life@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fire = "^0.1.3"
PyGithub = "^1.43"

[tool.poetry.dev-dependencies]
autopep8 = "^1.4"
flake8 = "^3.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
